S6L91 - Testing API
Found that this is not complete. He did not account for pulling items from a particular store, so if you try to add one to a store but it is in a different one it will not let you POST. And if you use update with PUT it looks like it clears out the other store ... Will need to fix later.